# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "summary": "Login and get tokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "Auth_refreshTokens",
        "summary": "Refresh access tokens",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "Auth_logout",
        "summary": "Logout and invalidate refresh tokens",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful logout"
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "Auth_getProfile",
        "summary": "Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile returned successfully"
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/branches": {
      "post": {
        "operationId": "Branches_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "Branches_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BranchResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/branches/{id}": {
      "get": {
        "operationId": "Branches_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "Branches_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "Branches_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/categories": {
      "post": {
        "operationId": "Categories_create",
        "summary": "Create a new category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "Categories_findAll",
        "summary": "Get all active categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "operationId": "Categories_findOne",
        "summary": "Get a category by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "Categories_update",
        "summary": "Update a category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "Categories_remove",
        "summary": "Soft delete a category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TeaP ERP/POS API",
    "description": "API documentation for TeaP — Bubble Tea Chain ERP/POS Management System. Supports multi-branch operations, POS, inventory management, HR, and finance.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT access token",
        "in": "header"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@teap.com"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "MANAGER",
              "HR",
              "ACCOUNTANT",
              "CASHIER",
              "WAREHOUSE_STAFF",
              "KITCHEN_STAFF"
            ]
          },
          "branchId": {
            "type": "string"
          },
          "allowedBranchIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "role",
          "allowedBranchIds"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "csrfToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfileDto"
          }
        },
        "required": [
          "accessToken",
          "csrfToken",
          "user"
        ]
      },
...
      "CreateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the branch"
          },
          "address": {
            "type": "string",
            "description": "The address of the branch"
          },
          "phone": {
            "type": "string",
            "description": "The phone number of the branch"
          }
        },
        "required": [
          "name",
          "address"
        ]
      },
      "BranchCountResponseDto": {
        "type": "object",
        "properties": {
          "users": {
            "type": "number"
          },
          "orders": {
            "type": "number"
          },
          "inventories": {
            "type": "number"
          }
        }
      },
      "BranchUserReferenceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "role"
        ]
      },
      "BranchResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "$ref": "#/components/schemas/BranchCountResponseDto"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchUserReferenceDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the branch"
          },
          "address": {
            "type": "string",
            "description": "The address of the branch"
          },
          "phone": {
            "type": "string",
            "description": "The phone number of the branch"
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "description": {
            "type": "string",
            "description": "Category description"
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order for display"
          }
        },
        "required": [
          "name"
        ]
      },
      "CategoryProductCountDto": {
        "type": "object",
        "properties": {
          "products": {
            "type": "number"
          }
        },
        "required": [
          "products"
        ]
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "$ref": "#/components/schemas/CategoryProductCountDto"
          }
        },
        "required": [
          "id",
          "name",
          "sortOrder",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "description": {
            "type": "string",
            "description": "Category description"
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order for display"
          }
        }
      },
...
    }
  }
}